Add skeleton of new capturer that uses Windows.Graphics.Capture API
This change lays the foundation for the new DesktopCapturer
implementation which will use the Windows.Graphics.Capture API.

In line with the other platform specific DesktopCapturer
implementations, I've moved the actual implementations into the win/
subdirectory and repurposed window_capturer_win.cc to instantiate
the most appropriate implementation. This will be where the WebRTC
field trial (or similar mechanism) and Windows version checks will go
when we begin to roll out the new implementation.

I've verified that the existing window capture functionality still works
by dropping these changes into the third_party/webrtc folder of a
Chromium enlistment, going to
https://webrtc.github.io/samples/src/content/getusermedia/getdisplaymedia/
and stepping through this new path under a debugger, and running the
existing WindowCapturerTests.

The next change in this series will begin to add functionality to the
new window_capturer_win_wgc files.

Bug: 9273
Change-Id: Ifc36ec69aed19563b9c20ef022760fb9c45cae25