Update patch set 4
Patch Set 4:

> Patch Set 4: Code-Review+1
> 
> > Patch Set 3:
> > 
> > Hi Taylor - I think it would make things simpler if we split the RTP implementation away from the SCTP (and sctp-like) code in the DataChannel and DataChannelController. As is, there's a lot of thread hopping happening and I the complexity of having all the data channel code mixed in these classes, makes any refactoring significantly more complex and risky.
> 
> Agreed. Do you have any other changes planned? What did you want to do in DataChannelController?

Plenty of stuff I'd _like_ to do 😊 but there are also several other things I'm working on, so I don't know exactly what I'll be able to do. I'll keep you in the loop though.

Here are some things that I'm thinkinf of right now (but have not started working on):
* Split DataChannel implementation into RtpDataChannel and DataChannel.
  - All state values including handshake states, would be managed fully
    in each class and ideally they would be fully independent of one
    another (i.e. not share any threading assumptions).
* Split DataChannelController into RTP and SCTP variants.
  - Again, separate implementations and I'd actually not want the two
    of them to share a parent class, since that will complicate further
    refactoring of either one.
* Change PeerConnection to have a pointer to either (xor) of the controller implementation.
  - Delay the construction of the controller class as much as possible.
  - It looks like PC gathers information about the state of the controllers bit by bit, so it takes
    a few steps for the controller instance to be fully constructed. This makes threading complicated
    and I'd like to be able to consider as much state as possible 'const'.
* Similarly for the SCTP data channel, if we can construct the DC instance itself as late as possible
  then there's possibly a way to construct it when an ID is available and that field can be
  considered const.
* Move logic relating to the data channels, that currently sits in the PC code, into the respective
  controller classes.
* Move construction of a DC instance into the controller classes.
  - effectively PC doesn't have to know if it has/constructs an RTP DC or a non-RTP DC.

So, basically address a few anti-patterns ("Cuckoo" in PC and trending a bit towards "God object") and trying to make more use of RAII.

After following how many thread hops happen during the course of sending or receiving a packet, I actually do think that it is likely contributing negatively to the performance of a DC object, including when it comes to throughput.

By the way, the pattern of having more than one implementation in one class (rtp vs not rtp for data channels and controllers), seems to also apply to PeerConnection. As an example, take a look at when we're using Unified Plan vs not and how long the Apply*Description methods have become.

With all of that taken care of, keeping the network state largely on the network thread, should allow us to remove synchronization and thread hops for the most common cases and when we do hop between threads, that that'll be for rare (comparatively) events.

Patch-set: 4