[Adaptation] Adaptation holds restrictions
This refactoring enables us to do multi-step adaptations
instead of the normal single step adaptations.

This work is required pre-requisite work to remove the
DropDueToSize methods from the ResourceAdaptationProcessorInterface.

This work also gives a path to fixing webrtc:11719, since double
adaptations can be done in a controlled manner with this API.


Bug: webrtc:11700,webrtc:11719
Change-Id: Ic7d27a49c82d0cf7ebaf79d7bdf0bc6f382770ed