[Adaptation] Remove degradation preference from VSE Resource Manager
Remove QP resource when QP scaling is disabled.

This removes the need for the QP scaler to spam underuse to make up for
the fact that the resource has been disabled. Now that the most limited
resource can be removed, the QP scaler's degradations will be removed
when the QP scaler is removed.

NOSUBMIT=true

Bug: webrtc:11563
Change-Id: I9dd7f874bb25301f2aceb11f8e9426f69d0c3148