red: only encode redundancy for active speech
When done unconditionally, RED will triple the bitrate. Using
a redundant encoding only if there is speech detected reduces
the bitrate cost of RED. A similar strategy is used for Opus
LBRR packets as described in
  https://tools.ietf.org/html/rfc6716#section-4.2.5
which says LBRR frames are only meant to be transmitted for active
speech. See also LBRR_SPEECH_ACTIVITY_THRES in libopus.

Note that it is not possible to encode redundancy with gaps as e.g.
[N, N-2] is ambigious and the receiver can not tell this apart from [N, N-1]

BUG=webrtc:11640,webrtc:11643

Change-Id: Idc0e79e351cc23e751f1b509331a75edc576616e