Begin implementing WGC CaptureFrame
This change introduces the design that will allow us to deliver frames
synchronously to callers despite the Windows.Graphics.Capture APIs being
inherently asynchronous.

We achieve this by having WindowCapturerWinWgc create and maintain a
WgcCaptureSession object for each window that it is asked to capture a frame
for. The capture session object will be the class that actually uses the WGC
APIs, and it will store the frames it receives in a frame pool and deliver them
via GetMostRecentFrame.

The next CL will add the necessary functionality to the WgcCaptureSession class.

Bug: webrtc:9273
Change-Id: I44e164f4874503d8ccc8e6a210e74f9c8458f6c4