Create tool to reencode old RTC event logs in the new format.
The most apparent different between the original and the reencoded copy is seen in the outgoing packet delay graphs where we seem to see random 1 ms changes. This is likely caused by the rounding from the old format which use us resolution to the new format which use millisecond resolution.

Bug: webrtc:11933
Change-Id: I806d762b28adfbec2b305b8b151981279c2eed59