AGC2 Saturation Protector: switch to ring buffer
Even if small, the peak delay buffer copies N-1 elements for each frame
whereas a ring buffer is copy-free and scales better if the buffer size
increases.

This CL also uses shorter names for the PeakEnveloper members and adds
a PeakEvenloper::Reset() to avoid calling a ctor when
SaturatioProtector::Reset() is called.

Bug: webrtc:7494
Change-Id: If8c33877b7ab1d881a0606e222b26857a82fff69