Add supported pixel formats property to encoder info
If the encoder supports the incoming pixel format then this frame should
be passed into the encoder without conversion if possible. That is, the
VideoStreamEncoder passes frames with the accepted types into the encoder without conversion.

Tested using video_loopback with NV12 generated frames for VP9, the only codec supporting NV12, as well as VP8 which only accepts I420 frames.

Bug: webrtc:11976,webrtc:11635
Change-Id: If39a815fb0c5636fceb1040c8946c3db2fb350a1