Make VideoFrameBuffer::GetNV12 virtual
Instead of checking the type, this is overridden by subclasses if they
or their underlying data are in NV12 format. This can be used by kNative
frames to expose underlying NV12 data in order to avoid extra conversions
to other pixel formats (ToI420).

The VideoStreamEncoder converts native frames using the GetXXX methods
available to first see if there are video frames underlying with underlying
data available. If not, it falls back to trying a conversion to I420.

Bug: webrtc:11977
Change-Id: I55b6ad4bc5a713e6026a030f0c25acb829b4e56f