Implement WgcCaptureSession
This change adds functionality to the WgcCaptureSession class, which is
used by the WindowCapturerWinWgc to capture application windows. Each
instance of WgcCaptureSession captures one window.

This change also adds necessary functionality to rtc_base/win, namely
HSTRING and GetActivationFactory functionality. These are required to
access WinRT APIs, such as the Windows.Graphics.Capture APIs.

The basic operation of a WgcCaptureSession is to receive an HWND
indicating the window to capture from the caller, and create a
GraphicsCaptureItem from it. We use this item to create a
Direct3D11CaptureFramePool and create and start a
GraphicsCaptureSession.

Once started, captured frames will begin to be deposited into the
FramePool. Typically, one would listen for the FrameArrived event and
process the frame then, but since we can't do anything with the frame
until the caller asks for one we ignore the FrameArrived event. Instead,
we wait for a request from the caller, then we get and process the
frame. This avoids a more complicated multi-threaded architecture.

Processing a frame involves moving the image data from an
ID3D11Texture2D stored in the GPU into a texture that is accessible
from the CPU, and then copying the data into the new DesktopFrameWinWgc
class. This copy is necessary as otherwise we would need to manage the
lifetimes of the CaptureFrame and ID3D11Texture2D objects, lest the
buffer be invalidated.

Once we've copied the data and passed it to the caller, we can unmap
the texture and exit the scope of the GetMostRecentFrame method, which
will destruct the CaptureFrame object. At this point, the CaptureSession
will begin capturing a new frame, and will soon deposit it into the
FramePool and we can repeat.

Bug: webrtc:9273
Change-Id: I8bb34669c72d30158052bb1a8ba4b41a5fe2b25c