Allow encoders to receive preferred pixel formats from native buffers
Adds a field to EncoderInfo called preferred_pixel_formats which a
software encoder populates with the pixel formats it supports. When a
kNative frame is received for encoding, the VideoStreamEncoder will
first try to get a frame in main memory in that pixel format from the
kNative frame, and fallback to converting the frame using ToI420.

This minimizes the number of conversions made in the case that the
encoder supports the pixel format of the native buffer or where conversion
can be accelerated. For example, in Chromium, the data backing a kNative
frame can either be in main memory in I420 or in gpu memory in NV12. This
method would allow the retreival of either an NV12 or I420 frame without
conversion or a copy.

Testing: Tested in Chrome with media::VideoFrame adapters.

Bug: webrtc:11977
Change-Id: I9becc4100136b0c0128f4fa06dedf9ee4dc62f37