Updates how min bitrate is set for VP9 SVC in perf tests.
In the video quality tests the codec-level min bitrate is sometimes set
as if single-stream simulcast will be used. When VP9 spatial layers are
then generated the will get new appropriate min bitrate levels.
The encoder adjuster can however look at the codec level min bitrate
and incorrectly adjust the bitrate up if it is set too high.

This CL removes the codec-level min bitrate if svc is used.

Bug: webrtc:12080
Change-Id: I563a57f3031c90c116448f1d255d3b6711f4ee75