Flush pending packets from BaseChannel when demux criteria changes.
When BaseChannel is changing its demux criteria (for example, removing
an SSRC), packets that were demuxed based on the old criteria may still
be in the worker thread's queue, due to the fact that packets are
asynchronously passed from a network thread to the worker thread.

If these packets are left to be delivered to the media channel, it will
possibly create a receive stream for an SSRC the media channel is no
longer supposed to handle, which could create problems down the road.

Fixed by using a separate AsyncInvoker for RTP packet delivery, and
flushing it when the demux criteria changes.k

Bug: b/172285886
Change-Id: I441a4c80daee9739b4c25c426537cef3ef887e0f