Use frame rate in video overhead calculation.
Assume bitrate is evenly distributed between frames. This is wrong for
uneven frame sizes and will underestimate for simulcast. However, it
will be more correct than the current calculation, especially for low
bitrates when each frame is smaller than one packet. This is also when
overhead matters more since it is a larger fraction of the total
bitrate.

Bug: b/166341943
Change-Id: I247b9d0fc7a8ad5daa9b577f55ec16c56efa34c3