Address CL comments from 200161.
This is a follow up to
https://webrtc-review.googlesource.com/c/src/+/200161

I forgot to actually upload the code that addressed reviewer's comments,
and since they were minor comments I went ahead and completed the CL
before I realized.

Fortunately no harm done because all of this code is behind a disabled
build flag. This is a great example of move fast, make mistakes. Lesson
learned.

Bug: webrtc:9273
Change-Id: I5e35b31719b264855568de4b5e595fe0f192654e