Make VideoFrameBuffer getters virtual.
Prior to this CL, VideoFrameBuffer assumed that if a VideoFrameBuffer
is of a particular type, e.g. NV12, then that class must implement the
corresponding type's buffer interface, i.e. NV12BufferInterface.

This is not always the case, for example WebRtcVideoFrameAdapter
implements VideoFrameBuffer even though it acts as an adapter for any
of the types. In the case of NV12, WebRtcVideoFrameAdapter creates an
internal NV12FrameAdapter that is returned, for example as a result of
CropAndScale(). However, having scaling be a prerequisite for GetNV12()
does not make a lot of sense.

By making GetXXX() virtual we allow overriding them to return internal
type-specific representations of the desired type.

This will also allow using CropAndScale() for soft-applying scaling and
GetXXX() to hard-apply them by overriding them in future CLs.

Bug: webrtc::12469
Change-Id: Id6968b7fe7bf9efb638452596795967b1fcc9d63